+++ R/show_ex_toc.R
@@ -20,14 +20,13 @@ show_ex_toc <- function(header = "", clear.it = TRUE) {
       options(learnitdown_ex_toc = NULL)
   }
 
-  # Also compile a single assignments.html file
+  # Also compile a single assignments.csv file
   ex_dir <- file.path(.get_output_dir(), "ex")
   files <- dir(ex_dir, pattern = "^assignment_.+\\.csv$", full.names = TRUE)
   res <- data.frame()
   for (file in files)
     res <- rbind(res, read.csv(file))
-  write.csv(res, file.path(ex_dir, "assignments.csv"))
-
+  write.csv(res, file.path(ex_dir, "assignments.csv"), row.names = FALSE)
   toc
 }